Create Actuator and Serialization classes. Consolidate OpenC2 Command dispatch functionality into the Consumer class.  Create sample actuator for SLPF. Add links to relevant OpenC2 documentation
The Actuator and Serialization classes serve to organize the implementation
of actuator-specific functions and methods for deserializing and serializing
OpenC2 messages. Instances of these classes can be added to a Consumer during
or after its initialization to extend its capabilities.
Add an example implementation of the SLPF actuator profile using the Actuator
class to the examples.